requirements.txt
discord.py==2.5.0
requests==2.32.3
pillow==11.1.0
